Modify ben to include sub-packages
