+++ src/main/java/ben/commands/AddCommand.java
@@ -1,4 +1,13 @@
-package ben;
+package ben.commands;
+
+import ben.exceptions.EmptyDescriptionException;
+import ben.exceptions.InvalidCommandException;
+import ben.storage.TaskList;
+import ben.tasks.Deadline;
+import ben.tasks.Event;
+import ben.tasks.Task;
+import ben.tasks.ToDo;
+import ben.ui.Ui;
 
 import java.time.LocalDateTime;
