+++ src/main/java/ben/commands/Command.java
@@ -1,4 +1,9 @@
-package ben;
+package ben.commands;
+
+import ben.exceptions.EmptyDescriptionException;
+import ben.exceptions.InvalidCommandException;
+import ben.storage.TaskList;
+import ben.ui.Ui;
 
 import java.time.format.DateTimeParseException;
