+++ src/main/java/ben/commands/FindCommand.java
@@ -1,4 +1,8 @@
-package ben;
+package ben.commands;
+
+import ben.storage.TaskList;
+import ben.exceptions.EmptyDescriptionException;
+import ben.ui.Ui;
 
 import java.util.Arrays;
