+++ src/main/java/ben/ui/MainWindow.java
@@ -1,5 +1,6 @@
-package ben;
+package ben.ui;
 
+import ben.Ben;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
