+++ src/main/java/ben/ui/Ui.java
@@ -1,4 +1,6 @@
-package ben;
+package ben.ui;
+
+import ben.storage.TaskList;
 
 import java.util.Scanner;
